requirements.txt
# Deep Learning & Detection
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0

# Numerical Computing
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
